--- python/photon/photon/__init__.py
@@ -1,6 +1,0 @@
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-from core.src.photon.libphoton import *
-from .photon_services import *
