--- src/plasma/plasma/__init__.py
@@ -1,5 +1,0 @@
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-from plasma.plasma import *
